Remove extra standard project and meanwhile disable behavior on .NET Framework
