--- src/System.Runtime.Serialization.Formatters/tests/NetStandardLib/Configurations.props
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <BuildConfigurations>
-      netstandard;
-    </BuildConfigurations>
-  </PropertyGroup>
-</Project>
